fixed null error psql
