+++ backend/src/main/java/com/musicme/musicme/controller/FileUploadController.java
@@ -1,0 +1,33 @@
+package com.musicme.musicme.controller;
+
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+@RestController
+public class FileUploadController {
+    @RequestMapping(value = "/upload",
+            method = RequestMethod.POST,
+            consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+
+    public String fileUpload(@RequestParam("file") MultipartFile file) throws IOException {
+        File convertFile = new File("/Users/kostas/Desktop/test" + file.getOriginalFilename());
+        convertFile.createNewFile();
+
+        try (FileOutputStream fout = new FileOutputStream(convertFile)) {
+            fout.write(file.getBytes());
+        } catch (Exception exe) {
+            exe.printStackTrace();
+        }
+        return "File has uploaded successfully";
+    }
+
+}
