+++ backend/src/main/resources/application.yml
@@ -14,6 +14,9 @@ spring:
       naming-strategy: org.hibernate.cfg.ImprovedNamingStrategy
     properties:
       hibernate:
+        jdbc:
+          lob:
+            non_contextual_creation: true
         dialect: org.hibernate.dialect.PostgreSQL9Dialect
   security:
     oauth2:
